text-to-speech/requirements.txt
marshmallow
redis
rx
wheel
orjson
google-cloud-texttospeech
s3fs